# Pertemuan 11/Restfull API/public/assets/api.yaml
openapi: 3.0.0
paths:
...
  /rivers:
    get:
      tags:
        - Rivers
      summary: 'Find list Rivers'
      description: 'Returns list of Rivers'
      operationId: getRivers
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Rivers' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Rivers' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/Rivers' } }
                type: object
        '404':
          description: 'Rivers not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - Rivers
      summary: 'Add a new Rivers to the store'
      operationId: addRivers
      requestBody:
        $ref: '#/components/requestBodies/Rivers'
      responses:
        '201':
          description: 'Created Rivers'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rivers'
            application/xml:
              schema:
                $ref: '#/components/schemas/Rivers'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/rivers/{id}':
    get:
      tags:
        - Rivers
      summary: 'Find Rivers by ID'
      description: 'Returns a single Rivers'
      operationId: getRiversById
      parameters:
        -
          name: id
          in: path
          description: 'ID of Rivers to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rivers'
            application/xml:
              schema:
                $ref: '#/components/schemas/Rivers'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'Rivers not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - Rivers
      summary: 'Update an existing Rivers'
      operationId: updateRivers
      parameters:
        -
          name: id
          in: path
          description: 'Rivers id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/Rivers'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Rivers not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - Rivers
      summary: 'Deletes a Rivers'
      operationId: deleteRivers
      parameters:
        -
          name: id
          in: path
          description: 'Rivers id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /supportDevices:
    get:
      tags:
        - SupportDevices
      summary: 'Find list SupportDevices'
      description: 'Returns list of SupportDevices'
      operationId: getSupportDevices
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SupportDevices' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SupportDevices' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/SupportDevices' } }
                type: object
        '404':
          description: 'SupportDevices not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - SupportDevices
      summary: 'Add a new SupportDevices to the store'
      operationId: addSupportDevices
      requestBody:
        $ref: '#/components/requestBodies/SupportDevices'
      responses:
        '201':
          description: 'Created SupportDevices'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportDevices'
            application/xml:
              schema:
                $ref: '#/components/schemas/SupportDevices'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/supportDevices/{id}':
    get:
      tags:
        - SupportDevices
      summary: 'Find SupportDevices by ID'
      description: 'Returns a single SupportDevices'
      operationId: getSupportDevicesById
      parameters:
        -
          name: id
          in: path
          description: 'ID of SupportDevices to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportDevices'
            application/xml:
              schema:
                $ref: '#/components/schemas/SupportDevices'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'SupportDevices not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - SupportDevices
      summary: 'Update an existing SupportDevices'
      operationId: updateSupportDevices
      parameters:
        -
          name: id
          in: path
          description: 'SupportDevices id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/SupportDevices'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'SupportDevices not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - SupportDevices
      summary: 'Deletes a SupportDevices'
      operationId: deleteSupportDevices
      parameters:
        -
          name: id
          in: path
          description: 'SupportDevices id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
...
    Rivers:
      title: Rivers
      description: Rivers
      properties:
        river_id:
          nullable: false
          title: river_id
          description: river_id
          type: string
          format: '-'
          maxLength: 6
        river_name:
          nullable: false
          title: river_name
          description: river_name
          type: string
          format: '-'
          maxLength: 50
        river_geojson:
          nullable: false
          title: river_geojson
          description: river_geojson
          type: string
          format: '-'
          maxLength: 100
        river_longitude:
          nullable: false
          title: river_longitude
          description: river_longitude
          type: string
          format: '-'
          maxLength: 100
        river_lattitude:
          nullable: false
          title: river_lattitude
          description: river_lattitude
          type: string
          format: '-'
          maxLength: 100
        created_at:
          nullable: false
          title: created_at
          description: created_at
          type: string
          format: '-'
        updated_at:
          nullable: false
          title: updated_at
          description: updated_at
          type: string
          format: '-'
      type: object
    SupportDevices:
      title: SupportDevices
      description: SupportDevices
      properties:
        support_device_id:
          nullable: false
          title: support_device_id
          description: support_device_id
          type: string
          format: '-'
          maxLength: 6
        regional_unit_id:
          nullable: false
          title: regional_unit_id
          description: regional_unit_id
          type: string
          format: '-'
          maxLength: 6
        support_device_name:
          nullable: false
          title: support_device_name
          description: support_device_name
          type: string
          format: '-'
          maxLength: 50
        support_device_description:
          nullable: false
          title: support_device_description
          description: support_device_description
          type: string
          format: '-'
        support_device_longitude:
          nullable: false
          title: support_device_longitude
          description: support_device_longitude
          type: string
          format: '-'
          maxLength: 100
        support_device_lattitude:
          nullable: false
          title: support_device_lattitude
          description: support_device_lattitude
          type: string
          format: '-'
          maxLength: 100
        support_device_type:
          nullable: false
          title: support_device_type
          description: support_device_type
          type: string
          format: '-'
        support_device_link:
          nullable: false
          title: support_device_link
          description: support_device_link
          type: string
          format: '-'
        support_device_status:
          nullable: false
          title: support_device_status
          description: support_device_status
          type: string
          format: '-'
        support_device_username:
          nullable: false
          title: support_device_username
          description: support_device_username
          type: string
          format: '-'
          maxLength: 255
        support_device_password:
          nullable: false
          title: support_device_password
          description: support_device_password
          type: string
          format: '-'
          maxLength: 150
        support_device_port:
          nullable: false
          title: support_device_port
          description: support_device_port
          type: string
          format: '-'
          maxLength: 20
        last_updated_condition_by:
          nullable: false
          title: last_updated_condition_by
          description: last_updated_condition_by
          type: string
          format: '-'
          maxLength: 6
        created_at:
          nullable: false
          title: created_at
          description: created_at
          type: string
          format: '-'
        updated_at:
          nullable: false
          title: updated_at
          description: updated_at
          type: string
          format: '-'
      type: object
...
  requestBodies:
...
    Rivers:
      description: 'Rivers object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/Rivers'
        application/xml:
          schema:
            $ref: '#/components/schemas/Rivers'
        application/json:
          schema:
            $ref: '#/components/schemas/Rivers'
    SupportDevices:
      description: 'SupportDevices object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/SupportDevices'
        application/xml:
          schema:
            $ref: '#/components/schemas/SupportDevices'
        application/json:
          schema:
            $ref: '#/components/schemas/SupportDevices'
...